+++ jest.config.js
@@ -5,4 +5,8 @@ module.exports = {
   collectCoverage: true,
   collectCoverageFrom: ["lib/**/*.ts", "!lib/**/*.d.ts"],
   coveragePathIgnorePatterns: ["/node_modules/", "/android/", "/ios/"],
+  moduleNameMapper: {
+    "^expo-file-system/next$":
+      "<rootDir>/__tests__/mocks/expo-file-system-next.ts",
+  },
 };